backend/requirements.txt
# Dependencias básicas para FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validación
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# Autenticación y seguridad  
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Variables de entorno
python-dotenv==1.0.0

# Cliente HTTP para health checks
httpx==0.25.1

# Upload de archivos e imágenes
pillow==10.1.0
azure-storage-blob==12.19.0
aiofiles==23.2.1

# Rate limiting y middleware
slowapi==0.1.9
redis==5.0.1

# Logging estructurado
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1